Changes to forms
Added a list of help texts for select and multiple select fields, to be
used as help on the front end.
Added FY and Q options for daterangepicker fields, which I expect to
implement as a group of formly fields. That means making the transition
to Angular from AngularJS. This should obsolete the fiscalyears and
quarters tables.
Made my FiscalQuarterField subclass DateIntervalField, instead of
replacing it.